refactor(viewer): modify getMovedPoints location to Annotation
