+++ libs/insight-viewer/src/Annotation/utils/getMovedPoints.spec.ts
