+++ libs/insight-viewer/src/Annotation/utils/getMovedPoints.ts
